neigh: Rework ARP handling to let the kernel do the resolution
Let the kernel do the work instead of open coding ARP resolution in the agent.
Use NTF_EXT_MANAGED iff possible from kernel side. This work was presented also
during this year's Plumbers conference:

  https://linuxplumbersconf.org/event/11/contributions/953/

And recently it has been merged into the kernel / iproute2:

  https://lore.kernel.org/netdev/20211011121238.25542-1-daniel@iogearbox.net/
  https://lore.kernel.org/netdev/20211013132140.11143-1-daniel@iogearbox.net/
  https://lore.kernel.org/netdev/20211025154728.2161-1-daniel@iogearbox.net/

NTF_EXT_MANAGED flagged entries are L3 addresses of next hops which we push
into the kernel without a L2 lladdress. These are being placed onto a managed
neighbor list which the kernel periodically walks and auto-refreshes such that
the entry is kept in REACHABLE state if possible. So aside from agent pushing
these down (and deleting them upon node delete events), there's no further
work required such as refresh controller, etc. Upon start, the agent probes
availability of the feature and sets option.Config.ARPPingKernelManaged when
possible to transparently opt into it.

For older kernels, we also reuse kernel infrastructure instead of open coding
an arping library by probing neighbor entries via NTF_USE flag. Similarly,
these entries we push down are L3 addresses of next hops without L2 lladdress.
The NTF_USE flag causes the kernel to trigger a one-time refresh of the entry
if it was in a STALE state. In contrast to NTF_EXT_MANAGED we cannot be fully
hands off in this case, since this still requires the agent's neighbor refresh
controller to trigger re-resolution. NTF_EXT_MANAGED avoids the extra netlink
churn when having lots of backends.

In either case, we push entries down via NTF_EXT_LEARNED flag such that these
are not GC'ed by the kernel. This is important since such neighbor entries are
used in fast-path, for example, to forward LB traffic to backends via XDP or
tc BPF programs, and at that layer it is not possible to resolve these on demand
through a slow-path. If kernel neighbor GC would evict them under pressure, then
next requests to such backends would fail. Thus, this preserves part of the old
NUD_PERMANENT properties, but for dynamically resolved entries.

Note that in contrast to NUD_PERMANENT entries, any active node traffic to such
neighbor from the networking stack will also update/refresh a potential STALE to
REACHABLE state. Also, removing the ARP library through this infrastructure in
here also has the advantage that we can now support IPv6 neighbor resolution with
the same mechanism. The agent is currently lacking this with the ARP library it
uses, so backends may not be able to get resolved properly. The IPv6 support
will be added separately and requires some more code refactoring.

Some more agent/implementation specific notes in terms of the ARP rework:

We still keep the next hop tracking in the agent in place since these need to be
reference counted there in case, for example, backend IP are used for multiple
services and they are direct next hops. We store neigh entries directly in the
agent, but with the transition we don't store neigh.HardwareAddr anymore. This
is okay as the way we update/delete them, they are not required, e.g. deletes
don't look at the lladdress in the kernel for a match.

For older kernels the NTF_EXT_LEARNED | NTF_USE handling is a bit cumbersome,
since the kernel has some behavioral quirks. We fixed them in above patches,
but most kernels don't have them at this point. Upon creation of new neighbor
entries, NTF_EXT_LEARNED flag is not being propagated, so we need to first
create a new entry e.g. in NUD_NONE pseudo-state with NTF_EXT_LEARNED and then
migrate this off from NUD_NONE with another request combined with NTF_USE to
trigger resolution. For already existing dynamic entries, this is not needed.

The same situation is when we migrate NUD_PERMANENT neighbor states over to
dynamic ones in NodeCleanNeighbors(). Older kernels don't have a direct path
to switch to NTF_EXT_LEARNED | NTF_USE, so we need to graduate them through
NUD_REACHABLE with NTF_EXT_LEARNED and then triggering another request with
NTF_USE flag. However, that migration is just a one-time event.

Related: #14962
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>