neigh: Fully rely on fib lookup for service handling
Given we use NTF_EXT_MANAGED, the neighbor entries in the kernel are auto-
refreshed and this also means there is no point in having the faster BPF
map as cache in front of it. If we go an optimize, then lets better do it
from kernel side for the neigh lookup rather than adding complexity from
BPF in front. Thus, this removes NODEPORT_NEIGH{4,6} maps with all related
surrounding bits which should also save quite a bit in terms of memory usage.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>